+++ src/app/components/account/account.component.html
@@ -106,7 +106,7 @@
     align-content: center;
     text-align: center;"
       >
-        {{this.multilinguism.translate('conçuPour')}}}
+        {{this.multilinguism.translate('conçuPour')}}
       </div>
     </div>
   </div>
