+++ src/app/components/header-print/exportStyle.component.css
@@ -26,3 +26,4 @@ textarea{
 #labelTextArea{
   transform: translateY(-50px);
 }
+
