+++ src/app/components/header-print/exportStyle.component.html
@@ -44,3 +44,41 @@
 
   </div>
 </div>
+
+<div class="footer width-100">
+  <div class="footer-page">
+
+    <div class="footer-item">
+      <label for="enableFooter">{{this.multilinguism.translate("enableFooter")}}</label>
+      <input type="checkbox" id="enableFooter" name="enableFooter" (click)="enableFooter()" [checked]="buttonEnableFooter">
+    </div>
+
+    <div class="footer-item" *ngIf="buttonEnableFooter">
+      <label for="btnText">{{this.multilinguism.translate("text")}}</label>
+      <input type="radio" id="btnText" name="typeChoiceFooter" (click)="choiceType('text')" [checked]="textButton">
+      <label class="space" for="btnImg">{{this.multilinguism.translate("image")}}</label>
+      <input type="radio" id="btnImg" name="typeChoiceFooter" (click)="choiceType('img')" [checked]="imgButton">
+    </div>
+
+    <div class="footer-item" *ngIf="typeChoice == 'text' && buttonEnableFooter">
+      <label id="labelTextArea" for="textArea">{{this.multilinguism.translate("textFooter")}}</label>
+      <textarea id="textArea" name="textFooter" rows="4" cols="50" (keyup)="getText($event)">{{this.footer}}</textarea>
+    </div>
+
+    <div class="footer-item" *ngIf="typeChoice == 'img' && buttonEnableFooter">
+      <label>{{this.multilinguism.translate("imageFooter")}}</label>
+      <input type="file" accept="image/*" (change)="onFileSelected($event)" [value]="selectedFileFooter">
+    </div>
+
+    <div class="footer-item" *ngIf="buttonEnableFooter">
+      <label for="positionFooter">{{this.multilinguism.translate("choosePosition")}}</label>
+      <select #option id="positionFooter" class="variantListBox" (input)="getPosition(option.value)" [(ngModel)]="textPositionFooter">
+        <option value="left">{{this.multilinguism.translate("leftPosition")}}</option>
+        <option value="center">{{this.multilinguism.translate("middlePosition")}}</option>
+        <option value="right">{{this.multilinguism.translate("rightPosition")}}</option>
+      </select>
+    </div>
+
+  </div>
+</div>
+
