+++ src/app/components/share/share.component.html
@@ -42,66 +42,5 @@
     <!--            </label>-->
     <!--            <input #zip type="file" (change)="exploreZip(zip.files)">-->
     <!--          </div>-->
-
-    <div class="export listElement">
-      <button mat-raised-button (click)="openDialogExportPages()">{{this.multilinguism.translate("exportPage")}}</button>
-    </div>
-
-    <div class="export listElement">
-      <label>
-        {{this.multilinguism.translate("exportPageAndItsSubset") + ":"}}
-      </label>
-      <select (change)="exportPageWithSubset()" class="selectBox" [(ngModel)]="pageIDToExport"
-              [ngModelOptions]="{standalone: true}">
-        <ng-container *ngFor="let p of this.boardService.board.PageList">
-          <option value="{{p.ID}}"> {{p.ID}} </option>
-        </ng-container>
-      </select>
-    </div>
-
-    <div class="export listElement">
-      <label>
-        {{this.multilinguism.translate("exportSave") + ":"}}
-      </label>
-      <input (click)="export()" type="button" value="Export">
-    </div>
-
-    <div class="export listElement">
-      <label>
-        {{this.multilinguism.translate("exportUser") + ":"}}
-      </label>
-      <input (click)="exportUser()" type="button" value="Export">
-    </div>
-
-    <div class="exportTo listElement">
-      <label>
-        {{this.multilinguism.translate("exportPDF") + ":"}}
-      </label>
-      <input (click)="this.printToPDF()" type="button" value="Export">
-    </div>
-
-    <div class="export listElement">
-      <label>
-        {{this.multilinguism.translate("exportTreeStructureAugCom") + ":"}}
-      </label>
-      <input (click)="exportTreeStructure()" type="button" value="Export">
-    </div>
-
-    <!--      <div>-->
-    <!--        <label class="floatLeft">-->
-    <!--          {{this.multilinguism.translate("importS4Y") + ":"}}-->
-    <!--        </label>-->
-    <!--        <input type="button" value="Import" (click)="parseAndCreateSpeak4YourselfGrid()">-->
-    <!--      </div>-->
-
-    <!--      <div>-->
-    <!--        <label class="floatLeft">-->
-    <!--          {{this.multilinguism.translate("importProloquo") + ":"}}-->
-    <!--        </label>-->
-    <!--        <input type="button" value="Import" (click)="parseAndCreateProloquoGrid()">-->
-    <!--      </div>-->
-
-    <!--    <app-saves></app-saves>-->
-
   </div>
 </div>
