+++ src/app/components/usertoolbar/usertoolbar.component.html
@@ -27,7 +27,7 @@
   <div class="button floatRight" id="lockUnlockButton" (click)="this.setLock()"
     [ngStyle]="{'margin-left': marginSize + 'px', 'margin-right' : '10px'}">
     <span class="tooltiptop">{{this.userToolBarService.unlock ?
-      this.multilinguism.translate('unlock'):this.multilinguism.translate('lock')}}</span>
+      this.multilinguism.translate('lock'):this.multilinguism.translate('unlock')}}</span>
     <div class="image width-100" [class.imageSelected]="this.userToolBarService.unlock"
       [ngStyle]="{'background-image':this.getIcon(this.userToolBarService.unlock?'unlock':'lock')}">
     </div>
