+++ src/assets/multilinguism.json
@@ -12,8 +12,8 @@
     { "id" : "paletteManagement" , "FR" : "Gestion des palettes", "EN" : "Palette Management" },
     { "id" : "language" , "FR" : "Langue", "EN" : "Language" },
     { "id" : "grammar" , "FR" : "Grammaire", "EN" : "Grammar" },
-    { "id" : "share" , "FR" : "Partager", "EN" : "Share" },
-    { "id" : "complementaryInfo" , "FR" : "Informations complémentaires", "EN" : "Complementary Information" },
+    { "id" : "share" , "FR" : "Importer", "EN" : "Import" }, 
+    { "id" : "complementaryInfo" , "FR" : "Informations complémentaires", "EN" : "Complementary Informations" },
     { "id" : "actuality" , "FR" : "Actualités", "EN" : "News" },
     { "id" : "contacts" , "FR" : "Contacts", "EN" : "Contacts" },
     { "id" : "deletePicture" , "FR" : "Supprimer la photo", "EN" : "Delete picture" },
@@ -93,13 +93,13 @@
     { "id" : "importZip" , "FR" : "Importer un fichier ZIP", "EN" : "Import ZIP" },
     { "id" : "exportSave" , "FR" : "Exporter une sauvegarde", "EN" : "Export save" },
     { "id" : "exportPDF" , "FR" : "Imprimer/Exporter en PDF", "EN" : "Print/export to PDF" },
-    { "id" : "exportTreeStructureAugCom" , "FR" : "Exporter AugCom arborescence", "EN" : "Export AugCom tree structure" },
+    { "id" : "exportTreeStructureAugCom" , "FR" : "Exporter vers Microsoft Word/Libre Office", "EN" : "Export to Microsoft Word/Libre Office" },
     { "id" : "exportUser", "FR": "Exporter cet utilisateur", "EN": "Export this user"},
     { "id" : "importS4Y" , "FR" : "Importer Speak4Yourself", "EN" : "Import Speak4Yourself" },
     { "id" : "importProloquo" , "FR" : "Importer Proloquo", "EN" : "Import Proloquo" },
     { "id" : "function" , "FR" : "Fonction", "EN" : "Function" },
     { "id" : "chooseFunction",  "FR" : "Choisissez une fonction", "EN" : "Choose a function" },
-    { "id" : "interactions",  "FR" : "Interactions", "EN" : "Interactions" },
+    { "id" : "interactions",  "FR" : "Sélection par fixation", "EN" : "Selection by focus" },
     { "id" : "otherForms",  "FR" : "Formes Alternatives", "EN" : "Alternative Forms" },
     { "id" : "information",  "FR" : "Informations", "EN" : "Informations" },
     { "id" : "applicationLanguage",  "FR" : "Langue de l'application", "EN" : "App Language" },
@@ -161,7 +161,7 @@
     { "id" : "VoulezVousVraimentSupprimerLaGrille?", "FR" : "Voulez vous vraiment supprimer la grille?", "EN" : "Do you really want to delete the grid?"},
     { "id" : "clearGrid", "FR" : "Vider la grille", "EN" : "Clear the grid"},
     { "id" :  "validate", "FR" : "Valider", "EN" :  "Validate"},
-    { "id":  "Print", "FR" :  "Impression", "EN" :  "Print"},
+    { "id":  "Print", "FR" :  "Exporter", "EN" :  "Export"},
     { "id":  "Header", "FR" :  "En-tête", "EN" :  "Header"},
     { "id":  "Footer", "FR" :  "Pied de page", "EN" :  "Footer"},
     { "id":  "enableHeader", "FR" :  "Activer l'en-tête", "EN" :  "Enable header "},
@@ -361,6 +361,8 @@
     { "id" :  "deactivateDwell", "FR": "Désactiver la sélection par focus", "EN": "Disable selection by focus"},
     { "id" : "buttonDescription", "FR": "L'élément créé sera un simple bouton", "EN": "The element will be a simple button"},
     { "id" : "folderDescription", "FR": "L'élément créé sera un dossier pouvant contenir d'autres boutons", "EN": "The element will be a folder that can contain other buttons"},
-    { "id" : "warningQuit", "FR": "Voulez vous vraiment quitter ? Les modifications non sauvegardées seront perdues", "EN": "Are you sure you want to quit ? Unsaved changes will be lost"}
+    { "id" : "warningQuit", "FR": "Voulez vous vraiment quitter ? Les modifications non sauvegardées seront perdues", "EN": "Are you sure you want to quit ? Unsaved changes will be lost"},
+    { "id" : "export", "FR": "Export", "EN": "Export"},
+    { "id" : "exportStyle", "FR": "Mise en page", "EN" : "Export style"}
   ]
 }
